backend/app/requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.1
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
jwt==1.3.1
psycopg==3.1.17
psycopg-binary==3.1.17
pycparser==2.21
PyJWT==1.7.1
pytz==2023.3.post1
slugify==0.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
